requirements.txt
asgiref==3.2.3
Django==3.0.3
django-elasticsearch-dsl==7.1.1
django-seed==0.2.2
django-taggit==1.2.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
Faker==4.0.1
hiredis==1.0.1
Markdown==3.2.1
model-bakery==1.1.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
redisearch==0.9.0
rmtest==0.7.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.8
